requirements.txt
requests==2.28.2
aiohttp==3.8.4
dnspython==2.3.0
python-nmap==0.7.1
beautifulsoup4==4.11.2
PyYAML==6.0
colorama==0.4.6
shodan==1.28.0
waybackpy==3.0.6
python-wappalyzer==0.3.1
tldextract==3.4.0
termcolor==2.2.0
aiodns==3.0.0
urllib3==1.26.14
pysocks==1.7.1
tqdm==4.65.0
cryptography==39.0.1
lxml==4.9.2
validators==0.20.0
tabulate==0.9.0
concurrent-log-handler==0.9.20 